+++ changelog.md
@@ -1,4 +1,7 @@
-## 26.1.2.5
+## 26.1.2.6
 
 - Added 16 ancient creatures: `Stegosaurus`, `Parasaurolophus`, `Spinosaurus`, `Carnotaurus`, `Velociraptor`, `Dilophosaurus`, `Argentinosaurus`, `Quetzalcoatlus`, `Mosasaurus`, `Plesiosaurus`, `Dunkleosteus`, `Smilodon`, `Woolly Mammoth`, `Woolly Rhinoceros`, `Dire Wolf`, and `Arthropleura`.
 - Added loot table for each species (Will be changed in future)
+- Removed species identification status screen in Identification Chamber
+- Added new keybind to open Species Journal (default key is `J`)
+- Added owner-only riding and movement controls for adult creatures.